Revert "MSS Live playing :"
This reverts commit 7bdf3a9f3d08c3f3633bd8eeb323a721b558dc93.
