--- src/mss/MssFragmentMoofProcessor.js
@@ -1,133 +1,0 @@
-/**
- * The copyright in this software is being made available under the BSD License,
- * included below. This software may be subject to other third party and contributor
- * rights, including patent rights, and no such rights are granted under this license.
- *
- * Copyright (c) 2013, Dash Industry Forum.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- *  * Redistributions of source code must retain the above copyright notice, this
- *  list of conditions and the following disclaimer.
- *  * Redistributions in binary form must reproduce the above copyright notice,
- *  this list of conditions and the following disclaimer in the documentation and/or
- *  other materials provided with the distribution.
- *  * Neither the name of Dash Industry Forum nor the names of its
- *  contributors may be used to endorse or promote products derived from this software
- *  without specific prior written permission.
- *
- *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS AS IS AND ANY
- *  EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- *  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
- *  IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
- *  INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
- *  NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
- *  PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
- *  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- *  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- *  POSSIBILITY OF SUCH DAMAGE.
- */
-
-import FactoryMaker from '../core/FactoryMaker';
-import ISOBoxer from 'codem-isoboxer';
-
-function MssFragmentMoofProcessor() {
-
-    let instance;
-
-    function processMoof(e) {
-
-        let i;
-        // e.request contains request description object
-        // e.response contains fragment bytes
-        let isoFile = ISOBoxer.parseBuffer(e.response);
-        // Update track_Id in tfhd box
-        let tfhd = isoFile.fetch('tfhd');
-        tfhd.track_ID = e.request.mediaInfo.index + 1;
-
-        // Add tfdt box
-        let tfdt = isoFile.fetch('tfdt');
-        let traf = isoFile.fetch('traf');
-        if (tfdt === null) {
-            tfdt = ISOBoxer.createFullBox('tfdt', traf, tfhd);
-            tfdt.version = 1;
-            tfdt.flags = 0;
-            tfdt.baseMediaDecodeTime = Math.floor(e.request.startTime * e.request.timescale);
-        }
-
-        let trun = isoFile.fetch('trun');
-
-        // Process tfxd boxes
-        // This box provide absolute timestamp but we take the segment start time for tfdt
-        let tfxd = isoFile.fetch('tfxd');
-        if (tfxd) {
-            tfxd._parent.boxes.splice(tfxd._parent.boxes.indexOf(tfxd), 1);
-            tfxd = null;
-        }
-        let tfrf = isoFile.fetch('tfrf');
-        if (tfrf) {
-            tfrf._parent.boxes.splice(tfrf._parent.boxes.indexOf(tfrf), 1);
-            tfrf = null;
-        }
-
-        // If protected content in PIFF1.1 format (sepiff box = Sample Encryption PIFF)
-        // => convert sepiff box it into a senc box
-        // => create saio and saiz boxes (if not already present)
-        let sepiff = isoFile.fetch('sepiff');
-        if (sepiff !== null) {
-            sepiff.type = 'senc';
-            sepiff.usertype = undefined;
-
-            let saio = isoFile.fetch('saio');
-            if (saio === null) {
-                // Create Sample Auxiliary Information Offsets Box box (saio)
-                saio = ISOBoxer.createFullBox('saio', traf);
-                saio.version = 0;
-                saio.flags = 0;
-                saio.entry_count = 1;
-                saio.offset = [];
-
-                let saiz = ISOBoxer.createFullBox('saiz', traf);
-                saiz.version = 0;
-                saiz.flags = 0;
-                saiz.sample_count = sepiff.sample_count;
-                saiz.default_sample_info_size = 0;
-                saiz.sample_info_size = [];
-
-                if (sepiff.flags & 0x02) {
-                    // Sub-sample encryption => set sample_info_size for each sample
-                    for (i = 0; i < sepiff.sample_count; i += 1) {
-                        // 10 = 8 (InitializationVector field size) + 2 (subsample_count field size)
-                        // 6 = 2 (BytesOfClearData field size) + 4 (BytesOfEncryptedData field size)
-                        saiz.sample_info_size[i] = 10 + (6 * sepiff.entry[i].NumberOfEntries);
-                    }
-                } else {
-                    // No sub-sample encryption => set default sample_info_size = InitializationVector field size (8)
-                    saiz.default_sample_info_size = 8;
-                }
-            }
-        }
-
-        tfhd.flags &= 16777214; // set tfhd.base-data-offset-present to false
-        tfhd.flags |= 131072; // set tfhd.default-base-is-moof to true
-        trun.flags |= 1; // set trun.data-offset-present to true
-
-        let moof = isoFile.fetch('moof');
-
-        let length = moof.getLength();
-        //offset is equal to length of the moof box + size and type definition for mdat box.
-        trun.data_offset = length + 8;
-
-        e.response = isoFile.write();
-    }
-
-    instance = {
-        processMoof: processMoof
-    };
-
-    return instance;
-}
-
-MssFragmentMoofProcessor.__dashjs_factory_name = 'MssFragmentMoofProcessor';
-export default FactoryMaker.getClassFactory(MssFragmentMoofProcessor);
